refactor: align types w hash_table and my_malloc, conform all testing to my_malloc, update Makefile
